+++ src/component/container/ErrorSnackbarContainer.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import ErrorSnackbar from '../presentational/ErrorSnackbar';
-import { hideErrorMessage } from '../../index';
+import { hideErrorMessage } from '../../actions';
 
 const mapStateToProps = state => ({
   open: state.error.open,
